requirements.txt
aiogram==2.25.1
openai==1.35.10
aiohttp==3.9.5
pytesseract==0.3.10
Pillow==10.4.0
pillow-heif==0.15.0
PyMuPDF==1.24.9
rapidfuzz==3.9.6
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
